Break sanitizers down to handler processors
