+++ xsdata/codegen/handlers/__init__.py
@@ -1,19 +1,41 @@
+from xsdata.codegen.handlers.attribute_compound_choice import (
+    AttributeCompoundChoiceHandler,
+)
+from xsdata.codegen.handlers.attribute_default_validate import (
+    AttributeDefaultValidateHandler,
+)
+from xsdata.codegen.handlers.attribute_default_value import AttributeDefaultValueHandler
+from xsdata.codegen.handlers.attribute_effective_choice import (
+    AttributeEffectiveChoiceHandler,
+)
 from xsdata.codegen.handlers.attribute_group import AttributeGroupHandler
 from xsdata.codegen.handlers.attribute_merge import AttributeMergeHandler
 from xsdata.codegen.handlers.attribute_mixed_content import AttributeMixedContentHandler
-from xsdata.codegen.handlers.attribute_sanitizer import AttributeSanitizerHandler
+from xsdata.codegen.handlers.attribute_name_conflict import AttributeNameConflictHandler
+from xsdata.codegen.handlers.attribute_overrides import AttributeOverridesHandler
+from xsdata.codegen.handlers.attribute_restrictions import AttributeRestrictionsHandler
 from xsdata.codegen.handlers.attribute_substitution import AttributeSubstitutionHandler
 from xsdata.codegen.handlers.attribute_type import AttributeTypeHandler
 from xsdata.codegen.handlers.class_enumeration import ClassEnumerationHandler
 from xsdata.codegen.handlers.class_extension import ClassExtensionHandler
+from xsdata.codegen.handlers.class_name_conflict import ClassNameConflictHandler
+
 
 __all__ = [
     "ClassEnumerationHandler",
     "AttributeGroupHandler",
     "AttributeMergeHandler",
-    "AttributeSanitizerHandler",
+    "AttributeOverridesHandler",
     "AttributeMixedContentHandler",
     "AttributeSubstitutionHandler",
     "AttributeTypeHandler",
     "ClassExtensionHandler",
+    "AttributeCompoundChoiceHandler",
+    "AttributeRestrictionsHandler",
+    "AttributeDefaultValueHandler",
+    "AttributeRestrictionsHandler",
+    "AttributeNameConflictHandler",
+    "AttributeDefaultValidateHandler",
+    "AttributeEffectiveChoiceHandler",
+    "ClassNameConflictHandler",
 ]
